+++ src/legacy/FakeLib/SquirrelHelper.fs
@@ -1,5 +1,5 @@
 /// Contains types and utility functions related to creating [Squirrel](https://github.com/Squirrel/Squirrel.Windows) installer.
-[<System.Obsolete("This API is obsolete. There is no alternative in FAKE 5 yet. You can help by porting this module.")>]
+[<System.Obsolete("FAKE0001 Use the Fake.Installer.Squirrel module instead")>]
 module Fake.Squirrel
 
 open Fake
@@ -11,7 +11,7 @@ open System.Text
 ///
 /// For reference, see: [Squirrel Command Line Options](https://github.com/Squirrel/Squirrel.Windows/blob/master/docs/advanced-releasify.md)
 [<CLIMutable>]
-[<System.Obsolete("This API is obsolete. There is no alternative in FAKE 5 yet. You can help by porting this module.")>]
+[<System.Obsolete("FAKE0001 Use the Fake.Installer.Squirrel module instead")>]
 type SquirrelParams =
     {
         /// The output directory for the generated installer
@@ -62,7 +62,7 @@ type SquirrelParams =
 /// - `SignExecutable` - `None`
 /// - `SigningKeyFile` - `None`
 /// - `SigningSecret` - `None`
-[<System.Obsolete("This API is obsolete. There is no alternative in FAKE 5 yet. You can help by porting this module.")>]
+[<System.Obsolete("FAKE0001 Use the Fake.Installer.Squirrel module instead")>]
 let SquirrelDefaults =
     let toolname = "Squirrel.exe"
     {
@@ -125,7 +125,7 @@ module internal ResultHandling =
 ///     Target "CreatePackage" (fun _ ->
 ///         SquirrelPack (fun p -> { p with WorkingDir = Some "./tmp" }) "./my.nupkg"
 ///     )
-[<System.Obsolete("This API is obsolete. There is no alternative in FAKE 5 yet. You can help by porting this module.")>]
+[<System.Obsolete("FAKE0001 Use Fake.Installer.Squirrel.releasify instead")>]
 let SquirrelPack setParams nugetPackage =
     use __ = traceStartTaskUsing "Squirrel" ""
     let parameters = SquirrelDefaults |> setParams